backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
ultralytics==8.3.0
Pillow==10.4.0
aiosqlite==0.20.0
PyJWT==2.9.0
